requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.0
fpdf==1.7.2
python-dotenv==1.0.0
